requirements.txt
aiohttp==3.10.5
altair==5.4.1
faiss-cpu==1.7.4
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
numpy==1.26.4
openai==1.44.0
pydantic==2.9.0
streamlit==1.38.0
tiktoken==0.7.0
upstash_redis